requirements.txt
crewai>=0.30.0
crewai[tools]>=0.30.0
litellm>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.5.3
requests>=2.31.0
httpx>=0.26.0
cerebras-cloud-sdk>=1.0.0
beautifulsoup4>=4.12.0
pytesseract>=0.3.10
Pillow>=10.0.0
pytest>=7.4.0

